Modularise the db classes
